Rename sentiment_feature.analyze() -> sentiment_merge.merge()
